pyproject.toml
[build-system]
requires = ["poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "Bluewind"
version = "0.0.1"
description = "Bluewind"
authors = ["Merwane Hamadi",]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "^3.8.1"
sqlparse = "^0.3.1"
python-dotenv = "^1.0.1"
django-environ = "^0.11.2"
behave-django = "^1.4.0"
psycopg2 = "^2.9.9"
gunicorn = "^22.0.0"
whitenoise = "^6.7.0"
pydeps = "^1.12.20"
python3-openid = "^3.2.0"
requests = "^2.32.3"
pyjwt = "^2.9.0"
cryptography = "^43.0.0"
requests-oauthlib = "^2.0.0"
factory-boy = "^3.3.0"
django-debug-toolbar = "^4.4.6"
model-mommy = "^2.0.0"
dj-database-url = "^2.2.0"
django = "^5.1"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.140.0"
botocore = "^1.35.2"
boto3 = "^1.35.2"
aiohttp = "^3.10.5"
google-cloud-pubsub = "^2.23.0"
django-clone = "^5.3.3"
colorlog = "^6.8.2"
rich = "^13.8.0"
django-allauth = "^64.2.0"
django-allauth-ui = "^1.4.1"
django-widget-tweaks = "^1.5.0"
django-ses = {extras = ["events"], version = "^4.1.1"}
autoflake = "^2.3.1"
ruff = "^0.6.3"
django-extensions = "^3.2.3"
nanoid = "^2.0.0"
django-fernet-encrypted-fields = "^0.1.3"
click = "^8.1.7"
django-formtools = "^2.5.1"
django-json-widget = "^2.0.1"
django-reverse-admin = "^2.9.6"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
drf-standardized-errors = "^0.14.1"
pylint = "^3.2.7"
autopep8 = "^2.3.1"
slippers = "^0.6.2"
watchdog = "^5.0.2"

[tool.poetry.extras]
argon2 = ["argon2-cffi",]
bcrypt = ["bcrypt",]

[tool.poetry.scripts]
django-admin = "django.core.management:execute_from_command_line"

[tool.poetry.dependencies.tzdata]
version = "*"
markers = "sys_platform == 'win32'"

[tool.ruff]
# The extend-ignore has been moved to the [tool.ruff.lint] section

[tool.ruff.lint]
select = ["E", "F", "I", "F811"]
ignore = ["E501"]
extend-ignore = ["E402"]  # Moved here from [tool.ruff]

[tool.ruff.lint.isort]
known-first-party = ["bluewind"]
known-third-party = ["django"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pylint.main]
ignore-patterns = ["^\\.#"]

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "C0330",  # bad-continuation
    "C0326",  # bad-whitespace
    "W0611",  # unused-import
    "W0703",  # broad-except
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
    "C0411",  # wrong-import-order
    "W0511",  # fixme
    "W1203", # logging-fstring-interpolation
    "E1101", # no-member
]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.design]
max-parents = 15

[tool.autopep8]
max_line_length = 120
in-place = true
recursive = true
aggressive = 3
ignore-local-config = true
